Refactoring subtree_source
